Version 0.5.0b
Added address book
You can now:
 - add addresses to your address book from recent transactions.
 - choose an address from your address book when sending coins.
 - show the QR code of any address in your address book
 - more fun stuff with addresses

BitcoinSpinner now recognizes Bitcoin address URLs
 on the form:
  "bitcoin:miGuMc6qtVEKS6Pf1jKddaa81DeHjMzkpB"
 or 
  "bitcoin:miGuMc6qtVEKS6Pf1jKddaa81DeHjMzkpB?amount=1.00"
This means that if you embed an URL like this on your web-page:
<a href="bitcoin:miGuMc6qtVEKS6Pf1jKddaa81DeHjMzkpB?amount=1.00">Donate</a>
You can click on it to send coins when you browse using your android phone. 
This also means that another Android App can launch the Send Bitcoins page of BitcoinSpinner, which is a cool move in the direction of in-app payments.

This version also contains minor UI fixes and tweaks.