Turning off SSL certificate verification when obtaining ticker info from MtGox.
For some reason the SSL certificate used by MtGox gives a security warning on my android device. This is not the case for any other browser I tried. Disabling SSL certificate checks is not a big problem as it is only used for obtaining ticker updates for BTC conversion to other currencies.